requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
python-dotenv
beautifulsoup4
requests
fastapi
uvicorn
jinja2
python-multipart
pydantic>=2.0.0
